move max value to bst
