+++ README.md
@@ -24,7 +24,7 @@ Clone the repo to your machine. Begin a new virtual environment with Python 3 an
     - [Towers of Hanoi](./challenges/towers_of_hanoi)
     - [Fizz Buzz Tree](./challenges/fizzbuzz-tree)
     - [Breadth-first Traversal](./data_structures/binary_search_tree/breadth_first_traversal.py)
-    - [Find the Maximum Value in a Binary Tree](./challenges/find-maximum-value-binary-tree)
+    - [Find the Maximum Value in a Binary Tree](./data_structures/binary_search_tree/find_maximum_value_binary_tree.py)
     - [Print Level Order of K-ary Tree](./challenges/print-level-order)
     - [Find Matches in K-ary Tree](./challenges/find_matches)
     - [Repeated Word](./challenges/repeated_word)
